move class Model to model.ts
