+++ lib/shared/src/models/index.test.ts
@@ -2,7 +2,6 @@ import { afterEach, beforeEach, describe, expect, it } from 'vitest'
 import { mockAuthStatus } from '../auth/authStatus'
 import { AUTH_STATUS_FIXTURE_AUTHED, type AuthenticatedAuthStatus } from '../auth/types'
 import {
-    Model,
     type ModelCategory,
     type ModelTier,
     ModelsService,
@@ -14,6 +13,7 @@ import {
 import { DOTCOM_URL } from '../sourcegraph-api/environments'
 import { CHAT_INPUT_TOKEN_BUDGET, CHAT_OUTPUT_TOKEN_BUDGET } from '../token/constants'
 import { getDotComDefaultModels } from './dotcom'
+import { Model } from './model'
 import { ModelTag } from './tags'
 import { ModelUsage } from './types'
